Skip new company creation if CRONUS exists

+++ Scripts/CustomActions/Initialize-NAVCompany.ps1
@@ -3,17 +3,28 @@
     Load-DockerInstanceAdminTools -Session $Session
     Invoke-Command -Session $Session -ScriptBlock {
         param([string]$ServerInstance)
-        Write-Host "Initializing company 'My Test Company'..."
-        Get-NAVCompany -ServerInstance $ServerInstance | Remove-NAVCompany -ServerInstance $ServerInstance -Force
-        New-NAVCompany -ServerInstance $ServerInstance -CompanyName "My Test Company" -EvaluationCompany        
+        $CompanyName=(Get-NAVCompany -ServerInstance $ServerInstance).CompanyName
+        if ($CompanyName -notmatch "CRONUS") {
+            $CompanyName = "My Test Company"
+            Write-Host "Initializing company ${CompanyName}..."
+            Get-NAVCompany -ServerInstance $ServerInstance | Remove-NAVCompany -ServerInstance $ServerInstance -Force
+            New-NAVCompany -ServerInstance $ServerInstance -CompanyName $CompanyName -EvaluationCompany        
+        }
         New-NAVServerUser -WindowsAccount $env:USERNAME -ServerInstance $ServerInstance -ErrorAction SilentlyContinue
         New-NAVServerUserPermissionSet -WindowsAccount $env:USERNAME -PermissionSetId SUPER -ServerInstance $ServerInstance -ErrorAction SilentlyContinue
-        Invoke-NAVCodeunit -CompanyName "My Test Company" -CodeunitId 2 -Language en-US -ServerInstance $ServerInstance -Force            
+        Invoke-NAVCodeunit -CompanyName $CompanyName -CodeunitId 2 -Language en-US -ServerInstance $ServerInstance -Force            
     } -ArgumentList $BranchSettings.instanceName
 } else {    
     Load-InstanceAdminTools -SetupParameters $SetupParameters 
-    Write-Host "Initializing company 'My Test Company'..."
-    Get-NAVCompany -ServerInstance $ServerInstance | Remove-NAVCompany -ServerInstance $ServerInstance -Force
-    New-NAVCompany -ServerInstance $ServerInstance -CompanyName "My Test Company" -EvaluationCompany        
-    Invoke-NAVCodeunit -CompanyName "My Test Company" -CodeunitId 2 -Language en-US -ServerInstance $ServerInstance -Force                
+    $CompanyName=(Get-NAVCompany -ServerInstance $ServerInstance).CompanyName
+    if ($CompanyName -notmatch "CRONUS") {
+        $CompanyName = "My Test Company"
+        Write-Host "Initializing company ${CompanyName}..."
+        Get-NAVCompany -ServerInstance $ServerInstance | Remove-NAVCompany -ServerInstance $ServerInstance -Force
+        New-NAVCompany -ServerInstance $ServerInstance -CompanyName $CompanyName -EvaluationCompany        
+    }
+    New-NAVServerUser -WindowsAccount $env:USERNAME -ServerInstance $ServerInstance -ErrorAction SilentlyContinue
+    New-NAVServerUserPermissionSet -WindowsAccount $env:USERNAME -PermissionSetId SUPER -ServerInstance $ServerInstance -ErrorAction SilentlyContinue
+    Invoke-NAVCodeunit -CompanyName $CompanyName -CodeunitId 2 -Language en-US -ServerInstance $ServerInstance -Force            
+
 }
